🎉(conversations) bootstrap backend & frontend
This is the first commit which provides all the first stack for a
working chat.

This is a first implementation with:
 - Vercel SDK for the frontend part
 - OpenAI Agent SDK for the backend

The stack can use a local LLM with docker ot a remote one.

This implementation is more a draft, but it provides the project
structure.

All tests are working even if we lack a lot of them.