💩(frontend) add source display in chat
This is a bad implementation of source display, it only demonstrate
the use of the message source sent by the backend.

Needless to say the UI is really bad on this.

We could improve this by fetching icon and page title in the backend
and send it directly in the `providerMetadata` of the source.
See: LanguageModelV1Source