⚡️(pydantic) use stored history in agent requests
We were converting the whole message history from the frontend
on each call. We replace that by direct use of the history stored
in the proper format from the database.

This important to do so because the raw Pydantic messages may contain
more information than the ones displayed to the user (like when
doing RAG web search).