⬇️(markitdown) enforce old 0.0.2 version
Our base docker image (alpine) does not support magika dependency
because there is no wheel for onnxruntime for alpine...
It's complex to build and a bit hackish.

Switching to a slim (debian) base image, is a solution but trivy
scan detects CVEs. We need to check the CVEs don't regard our stack
but it's bit of work also.

For the time being, I pin the markitdown version to an old one...
This is sad, since a lot of bugs has been fixed.