✨(backend) allow use to stop conversation streaming
This introduces a new endpoint to allow user to stop the message
stream.

We currently use uWSGI workers, which will not automatically stop the streaming
when the request is cancelled by the client. Therefore, we need to
explicitly stop the streaming by calling the /stop-steaming endpoint.
When (if) we switch to Gunicorn with Uvicorn workers, this will not be necessary
as the Uvicorn workers will automatically stop the streaming when the request
is cancelled. BUT, we will then need to handle the streaming cancellation when
the user is simply offline and still waiting for a response and the conversation
to be updated with the result. So this endpoint will still be useful to be able
to detect the cancellation of the streaming versus the user being offline.