💥(agent) replace routing w/ tool calls
The first implementation was made because our models had not
tool calling capabilities, and the routing and context management
were quite bad.

Now we can use tools (yay) the whole code is switched to tool
calling.

This first implementation of summarize agent is quite bad, this
will be improved soon.