✨(langfuse) allow user to score messages from LLM (#61)
* ✨(langfuse) allow user to score messages from LLM

We store the trace ID in the message ID... this is not very clean
but it is the easiest way to consistently save the trace ID.

This adds the endpoint for the "thumb up"/"thumb down" calls.

* 💄(frontend) add the thumb up/down to score LLM messages

This enables the score button when a trace exists (ie the
user allows analytics and Langfuse is setup).

* ✨(langfuse) allow user to score messages from LLM

allow user to score messages from LLM

---------

Co-authored-by: Eléonore Voisin <elevoisin@gmail.com>
Co-authored-by: elvoisin <95469923+elvoisin@users.noreply.github.com>