🧱(storage) upload the user documents into S3
The user uploaded content (images/documents) are now directly
uploaded to S3, and a pre-signed URL is given to the LLM.

The code is a bit complicated because there are behaviors
from docs, other from drive, and in the end some parts are
not used.

This will need some cleanup.

Improvements:
 - Allow to upload PDF without vectorized storage (to let
   LLM manage them on the fly)
 - Allow to configure whether a model can support images/PDF
   or not