💄(chat) add code highlighting for LLM responses
This adds a plugin to nicely display the code in the conversation.
The plugin is rehype-pretty-code, with the `github-dark-dimmed`
theme.

We may select another theme, and later also add light and dark
themes:
 - https://shiki.style/themes#themes

For later we may add instruction for the LLM to manage
transformers: https://shiki.style/packages/transformers

We also considered using https://sugar-high.vercel.app/ which
is quite light and powerful, we might switch later if pretty-code
is too heavy...

In both cases, the themes can be fully overriden if we want to
match the application colors ;)