⚗️(summarize) move the system prompt to instruction
When the tool is called, the agent graph call the LLM again with
the tool response, and the instructions. I hope using instruction
here will provide better results.

The former way to add the summarize tool as output does not work
properly with Mistral:
- if the user ask for a summary, the tool is called and the
  result is returned directly
- then if there is another user request which does not trigger
  a tool: boom, there is a JSON encode error...

I was not able to understand why this happens, so for now, the
summarize tool is not an "output".